+++ src/components/Section.js
@@ -1,0 +1,17 @@
+export class Section {
+    constructor({data, renderer}, containerSelector) {
+        this._renderedItems = data;
+        this._renderer = renderer;
+        this._container = document.querySelector(containerSelector);
+    }
+
+    addItem(element, isFirst=false) {
+        isFirst? this._container.prepend(element) : this._container.append(element);
+    }
+
+    renderItems() {
+        this._renderedItems.forEach(item => {
+            this._renderer(item);
+        });
+    }
+}
