+++ src/components/UserInfo.js
@@ -37,21 +37,3 @@ export class UserInfo {
             });
     }
 }
-
-export class Section {
-    constructor({data, renderer}, containerSelector) {
-        this._renderedItems = data;
-        this._renderer = renderer;
-        this._container = document.querySelector(containerSelector);
-    }
-
-    addItem(element, isFirst=false) {
-        isFirst? this._container.prepend(element) : this._container.append(element);
-    }
-
-    renderItems() {
-        this._renderedItems.forEach(item => {
-            this._renderer(item);
-        });
-    }
-}
