+++ src/components/constants.js
@@ -48,4 +48,3 @@ const myConfiguration = {
     },
     id: undefined,
 }
-
